test-lib: replace GIT_BUILD_DIR with TEST_TARGET_DIRECTORY
In the original Git test framework, the variable for test target
directory is named as "GIT_BUILD_DIR". Replace it with a more general
name "TEST_TARGET_DIRECTORY" so that test-lib can be reused in other
projects. The directory "TEST_TARGET_DIRECTORY" will be added to PATH so
that the target binaries to be tested can be found in this path first.

The default value of "TEST_TARGET_DIRECTORY" is the parent directory of
the current test directory. If you want to use a different
"TEST_TARGET_DIRECTORY", define it via environment.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>