test-lib: use TEST_LIB_DIRECTORY to locate test-lib
In the Git project, test framework files such as "test-lib.sh" and
"test-lib-function.sh" are stored in the same directory as test suites.
To make the git test framework easy to reuse, it is best to separate the
test framework from the test case files.

We add a new variable TEST_LIB_DIRECTORY which points to the test-lib
directory. The value of TEST_LIB_DIRECTORY is deduced from the variable
"$BASH_SOURCE", which only works with bash and may not work with other
shells. If fail to detect TEST_LIB_DIRECTORY, user can provide it via
the environment.

For example, if we save test-lib framework files in a "lib" directory
inside the test directory, a test suite file in the test directory can
can source test-lib file using directive as below.

    . lib/test-lib.sh

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>