test-lib: remove git installation check from test-lib
When running the test suite (t0000-basic.sh) in the "test-example"
directory, will get the following error:

    ./test-lib.sh: line 115: /path/to/test-lib/git: No such file or
    directory

Remove git installation checks from test-lib, so test suites in
"test-example" won't fail with the above error.

Directives removed from test-lib are added to the extension directory
of the "test-example-in-git" directory, so the test suites there will
work just like the original test suite in the Git project.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>