test-lib: load test extensions from extension dir
User can extend test-lib framework by providing additional scripts in
an extension directory ("sharness.d/" in TEST_DIRECTORY, which was
introduced in Sharness project). Extensions in this directory will be
loaded before running test cases.

We will move the instructions in test-lib that are strongly related to
the Git project to the extension directory of the "test-example-in-git"
directory, so that the test suites in  both "test-example-in-git" and
"test-example" directories can be executed correctly.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>