test-lib: implement test-tool core sub commands in python
The original test-tool was implemented in C, and was used widely in
test-lib (test-lib.sh and test-lib-functions.sh). Since "test-tool" is
developed using the git-core API, it is difficult to separate it from
the git source code. Moreover, the "test-tool" implemented in C language
needs to be compiled before using test-lib, which is very inconvenient.

Rewrite the core "test-tool" commands in python, so that with a proper
python interpreter installed we can use test-tool command directly without
compiling it.

Users can easily extend the "test-tool" in the test directory by creating
a Python module named "test-tools" in the test directory. For example,
users can develop a "test-tools" module as below:

    test-tools/__init__.py
    test-tools/foo-bar.py

Then users can use "test-tool foo-bar ..." command in their test cases.
Command arguments will pass to the "Run()" methods of the module.

In order to test the functional usability of the standalone "test-lib"
repository, we can check out this repository into the "t/" directory of
the Git workspace to replace the "t/" directory of Git itself for
testing.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>