test-lib: remove dependency on GIT-BUILD-OPTIONS
When running test suite (t0000-basic.sh) in "test-example" directory
without git building environment, get the errors below:

    error: GIT-BUILD-OPTIONS missing (has Git been built?).

Remove instructions of loading and checking the GIT-BUILD-OPTIONS
variable in test-lib, and set default path for tools directory. So
test suites in "test-example" won't fail with the above error.

Directives removed from test-lib are added to the extension directory
of the "test-example-in-git" directory, so the test suites there will
work just like the original test suite in the Git project.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>