test-lib: not assume TEST_DIRECTORY ends in '/t'
If the test directory where test suites are located is not named with
"./t", running the test suite (t0000-basic.sh) will give the following
error:

    PANIC: Running in a /path/to/test-lib that doesn't end in '/t'?

We once assumed that TEST_DIRECTORY ended with "/t", this restriction is
lifted so that test suites can be placed in any directory specified by
the project which use test-lib as its default test framework.

Now we can checkout out this repository to a directory other than 't/'
inside the Git worktree and start testing.

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>