test-lib: perf: reference test-lib with TEST_LIB_DIRECTORY
The performance test suites are stored in the "perf" directory, and
variables TEST_DIRECTORY and TEST_OUTPUT_DIRECTORY are hard-coded in the
performance test library (i.e. perf-lib.sh), making it difficult to
separate the performance test suites and the test-lib framework.

This problem can be solved by making some changes to the performance
test suites. Include both "test-lib.sh" and "perf-lib.sh" in a
performance test suite so that we can use the variable
"TEST_LIB_DIRECTORY" in "perf-lib.sh".

Signed-off-by: Jiang Xin <zhiyou.jx@alibaba-inc.com>